Alguna vez Creíste
¿Alguna vez creíste que podías hacer algo inolvidable, que no importa que pasen los años?. Todo lo que hagas importa y va a traspasar el tiempo en el que vos vivas. No te desanimes, después de todo, queremos que el mundo sea un lugar mejor y más justo para todos, ¿verdad?

Have you ever believed that you could do something unforgettable, that no matter what the years go by? Everything you do matters and will transcend the time in which you live. Don't be discouraged, after all, we want the world to be a better and fairer place for everyone, right?